Refactored project structure: moved response classes from DTO package to separate response package as per architecture
