+++ HealthCareHub/src/main/java/nimblix/in/HealthCareHub/response/AdmissionHistoryResponse.java
@@ -1,4 +1,6 @@
-package nimblix.in.HealthCareHub.dto;
+package nimblix.in.HealthCareHub.response;
+
+
 
 import lombok.*;
 
@@ -28,3 +30,4 @@ public class AdmissionHistoryResponse {
     private Long hospitalId;
     private String hospitalName;
 }
+
